KVM: unmap pages from the iommu when slots are removed
CVE-2012-2121

BugLink: http://bugs.launchpad.net/bugs/987569

commit 32f6daad4651a748a58a3ab6da0611862175722f upstream.

We've been adding new mappings, but not destroying old mappings.
This can lead to a page leak as pages are pinned using
get_user_pages, but only unpinned with put_page if they still
exist in the memslots list on vm shutdown.  A memslot that is
destroyed while an iommu domain is enabled for the guest will
therefore result in an elevated page reference count that is
never cleared.

Additionally, without this fix, the iommu is only programmed
with the first translation for a gpa.  This can result in
peer-to-peer errors if a mapping is destroyed and replaced by a
new mapping at the same gpa as the iommu will still be pointing
to the original, pinned memory address.

Signed-off-by: Alex Williamson <alex.williamson@redhat.com>
Signed-off-by: Marcelo Tosatti <mtosatti@redhat.com>
[bwh: Backported to 2.6.32:
 - Adjust context
 - In __kvm_set_memory_region(), call kvm_iommu_unmap_pages()
   immediately before kvm_free_physmem_slot() which cleans up the old
   memory slot.  Make this dependent on CONFIG_DMAR, consistent with
   the use of kvm_iommu_map_pages().]
Signed-off-by: Luis Henriques <luis.henriques@canonical.com>
Signed-off-by: Tim Gardner <tim.gardner@canonical.com>